NimphProject/Nproject/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysql-connector-python==9.0.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
